// types/company.types.ts
export interface Company {
  id: string;
  firebaseUid: string | null;
  name: string;
  ruc: number | null;
  phoneNumber: string | null;
  description: string | null;
  industry: string | null;
  address: string | null;
  reference: string | null;
  email: string | null;
  owner_name: string | null;
  deletedAt?: Date | null;
  isActive: boolean | null;
  customId: string | null;
  district: string | null;
}